Small project refactoring
