--- homework-10/src/demo_actions/fetchUsers.js
@@ -1,5 +1,0 @@
-const { usersService } = require('../services');
-
-module.exports = async () => {
-    return usersService.fetch();
-};
