--- homework-10/src/demo_actions/index.js
@@ -1,11 +1,0 @@
-const createUsers = require('./createUsers');
-const fetchUsers = require('./fetchUsers');
-const updateUsers = require('./updateUsers');
-const removeUsers = require('./removeUsers');
-
-module.exports = {
-    createUsers,
-    fetchUsers,
-    updateUsers,
-    removeUsers,
-};
