--- homework-10/src/demo_actions/removeUsers.js
@@ -1,5 +1,0 @@
-const { usersService } = require('../services');
-
-module.exports = async userId => {
-    return usersService.delete({ id: userId });
-};
