--- homework-10/src/demo_actions/updateUsers.js
@@ -1,5 +1,0 @@
-const { usersService } = require('../services');
-
-module.exports = async userId => {
-    return usersService.update({ id: userId }, { login: 'Dan' });
-};
