+++ homework-10/src/knexInstance.js
@@ -1,22 +1,4 @@
-const knexInstance = require('knex')({
-    client: 'mysql',
-    version: '5.7',
+const knex = require('knex');
+const knexConfig = require('../knexfile');
 
-    connection: {
-        host: process.env.DB_HOST,
-        user: process.env.DB_USER,
-        password: process.env.DB_PASSWORD,
-        database: process.env.DB_NAME,
-    },
-
-    pool: {
-        min: 2,
-        max: 10,
-    },
-
-    migrations: {
-        tableName: 'migrations',
-    },
-});
-
-module.exports = knexInstance;
+module.exports = knex(knexConfig[process.env.NODE_ENV || 'development']);
