--- homework-10/src/services/index.js
@@ -1,5 +1,0 @@
-const usersService = require('./users');
-
-module.exports = {
-    usersService,
-};
