+++ homework-10/src/usersService.js
@@ -1,4 +1,4 @@
-const knexInstance = require('../knexInstance');
+const knexInstance = require('./knexInstance');
 
 const TABLE_NAME = 'users';
